+++ evosax/strategies/lm_ma_es.py
@@ -5,16 +5,23 @@ from typing import Tuple
 from ..strategy import Strategy
 
 
-class MA_ES(Strategy):
-    def __init__(self, num_dims: int, popsize: int, elite_ratio: float = 0.5):
-        """MA-ES (Bayer & Sendhoff, 2017)
-        Reference: https://www.honda-ri.de/pubs/pdf/3376.pdf
+class LM_MA_ES(Strategy):
+    def __init__(
+        self,
+        num_dims: int,
+        popsize: int,
+        elite_ratio: float = 0.5,
+        memory_size: int = 10,
+    ):
+        """Limited Memory MA-ES (Loshchilov et al., 2017)
+        Reference: https://arxiv.org/pdf/1705.06693.pdf
         """
         super().__init__(num_dims, popsize)
         assert 0 <= elite_ratio <= 1
         self.elite_ratio = elite_ratio
         self.elite_popsize = int(self.popsize * self.elite_ratio)
-        self.strategy_name = "MA_ES"
+        self.memory_size = memory_size
+        self.strategy_name = "LM_MA_ES"
 
     @property
     def params_strategy(self) -> chex.ArrayTree:
@@ -56,20 +63,30 @@ class MA_ES(Strategy):
         weights_truncated = weights.at[self.elite_popsize :].set(0)
 
         # lrate for cumulation of step-size control and rank-one update
-        c_sigma = (mu_eff + 2) / (self.num_dims + mu_eff + 5)
+        c_sigma = jnp.minimum(2 * self.popsize / self.num_dims, 1)
         d_sigma = (
             1
             + 2
             * jnp.maximum(0, jnp.sqrt((mu_eff - 1) / (self.num_dims + 1)) - 1)
             + c_sigma
         )
-
         chi_n = jnp.sqrt(self.num_dims) * (
             1.0
             - (1.0 / (4.0 * self.num_dims))
             + 1.0 / (21.0 * (self.num_dims ** 2))
         )
 
+        c_d = jnp.array(
+            [1 / (1.5 ** i * self.num_dims) for i in range(self.memory_size)]
+        )
+        c_c = jnp.array(
+            [
+                self.popsize / (4 ** i * self.num_dims)
+                for i in range(self.memory_size)
+            ]
+        )
+        c_c = jnp.minimum(c_c, 1.99)
+        mu_w = 1 / jnp.sum(weights_truncated ** 2)
         params = {
             "mu_eff": mu_eff,
             "c_1": c_1,
@@ -77,9 +94,13 @@ class MA_ES(Strategy):
             "c_m": 1.0,
             "c_sigma": c_sigma,
             "d_sigma": d_sigma,
+            "c_c": c_c,
+            "c_d": c_d,
             "chi_n": chi_n,
+            "weights": weights,
             "sigma_init": 1.0,
             "weights_truncated": weights_truncated,
+            "mu_w": mu_w,
             "init_min": 0.0,
             "init_max": 0.0,
         }
@@ -100,7 +121,7 @@ class MA_ES(Strategy):
             "p_sigma": jnp.zeros(self.num_dims),
             "sigma": params["sigma_init"],
             "mean": initialization,
-            "M": jnp.eye(self.num_dims),
+            "M": jnp.zeros((self.num_dims, self.memory_size)),
         }
         return state
 
@@ -115,6 +136,8 @@ class MA_ES(Strategy):
             state["M"],
             self.num_dims,
             self.popsize,
+            params["c_d"],
+            state["gen_counter"],
         )
         return x, state
 
@@ -133,9 +156,8 @@ class MA_ES(Strategy):
         mean, z_k = update_mean(
             state["mean"], state["sigma"], sorted_solutions, params
         )
-
         p_sigma, norm_p_sigma = update_p_sigma(z_k, state["p_sigma"], params)
-        M = update_M_matrix(state["M"], state["p_sigma"], z_k, params)
+        M = update_M_matrix(state["M"], z_k, params)
         sigma = update_sigma(state["sigma"], norm_p_sigma, params)
 
         state["mean"] = mean
@@ -174,25 +196,20 @@ def update_p_sigma(
 
 
 def update_M_matrix(
-    M: chex.Array, p_sigma: chex.Array, z_k: chex.Array, params: chex.ArrayTree
+    M: chex.Array, z_k: chex.Array, params: chex.ArrayTree
 ) -> chex.Array:
     """Update the M matrix."""
-    rank_one = jnp.outer(p_sigma, p_sigma)
-    rank_mu = jnp.sum(
-        jnp.array(
-            [
-                w * jnp.outer(z, z)
-                for w, z in zip(params["weights_truncated"], z_k)
-            ]
-        ),
+    weighted_elite = jnp.sum(
+        jnp.array([w * z for w, z in zip(params["weights_truncated"], z_k)]),
         axis=0,
     )
-    M_new = M @ (
-        jnp.eye(M.shape[0])
-        + params["c_1"] / 2 * (rank_one - jnp.eye(M.shape[0]))
-        + params["c_mu"] / 2 * (rank_mu - jnp.eye(M.shape[0]))
-    )
-    return M_new
+    # Loop over individual memory components - this could be vectorized!
+    for i in range(M.shape[1]):
+        new_m = (1 - params["c_c"][i]) * M[:, i] + jnp.sqrt(
+            params["mu_w"] * params["c_c"][i] * (2 - params["c_c"][i])
+        ) * weighted_elite
+        M = M.at[:, i].set(new_m)
+    return M
 
 
 def update_sigma(
@@ -213,10 +230,17 @@ def sample(
     M: chex.Array,
     n_dim: int,
     pop_size: int,
+    c_d: chex.Array,
+    gen_counter: int,
 ) -> chex.Array:
     """Jittable Gaussian Sample Helper."""
     z = jax.random.normal(rng, (n_dim, pop_size))  # ~ N(0, I)
-    y = M.dot(z)  # ~ N(0, C)
-    y = jnp.swapaxes(y, 1, 0)
-    x = mean + sigma * y  # ~ N(m, σ^2 C)
+    for j in range(M.shape[1]):
+        update_bool = gen_counter > j
+        new_z = (1 - c_d[j]) * z + (c_d[j] * M[:, j])[:, jnp.newaxis] * (
+            M[:, j][:, jnp.newaxis] * z
+        )
+        z = jax.lax.select(update_bool, new_z, z)
+    z = jnp.swapaxes(z, 1, 0)
+    x = mean + sigma * z  # ~ N(m, σ^2 C)
     return x